refactor: relocate and modify handle db functions
